Temporarily duplicate code to avoid circular import problem
